+++ CHANGELOG.md
@@ -8,6 +8,7 @@ All notable changes to `homebridge-config-ui-x` will be documented in this file.
 
 - fix german welcome translation typo
 - add spinners for loading data widget values
+- hide ssl settings when on hb raspbian image
 
 ### Other Changes